Update deployment for L1 Bridge w/ ChugSplash
